pyproject.toml
[project]
name = "technicalassessmententropiaai"
version = "0.1.0"
description = "Technical Assessment AI Agent Trainee"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "google-generativeai>=0.8.6",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.0",
    "pandas>=3.0.0",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
]
